+++ Composer/packages/client/src/App.tsx
@@ -18,13 +18,13 @@ import { main, sideBar, content, divider, globalNav, leftNavBottom, rightPanel, 
 import { resolveToBasePath } from './utils/fileUtil';
 import { ErrorBoundary } from './components/ErrorBoundary';
 import { RequireAuth } from './components/RequireAuth';
-import { AppUpdater } from './components/AppUpdater';
 import onboardingState from './utils/onboardingStorage';
 import { isElectron } from './utils/electronUtil';
 
 initializeIcons(undefined, { disableWarnings: true });
 
 const Onboarding = React.lazy(() => import('./Onboarding'));
+const AppUpdater = React.lazy(() => import('./components/AppUpdater').then(module => ({ default: module.AppUpdater })));
 
 // eslint-disable-next-line react/display-name
 const Content = forwardRef<HTMLDivElement>((props, ref) => <div css={content} {...props} ref={ref} />);
@@ -141,6 +141,8 @@ export const App: React.FC = () => {
 
   const mapNavItemTo = x => resolveToBasePath(BASEPATH, x);
 
+  const renderAppUpdater = isElectron();
+
   const openedDialogId = designPageLocation.dialogId || dialogs.find(({ isRoot }) => isRoot === true)?.id || 'Main';
   return (
     <Fragment>
@@ -212,7 +214,7 @@ export const App: React.FC = () => {
           </ErrorBoundary>
         </div>
         <Suspense fallback={<div />}>{!state.onboarding.complete && <Onboarding />}</Suspense>
-        {isElectron() && <AppUpdater />}
+        <Suspense fallback={<div />}>{renderAppUpdater && <AppUpdater />}</Suspense>
       </div>
     </Fragment>
   );
